fix: remove GetDownScripts
