+++ packages/data-migration/src/methods/index.ts
@@ -1,6 +1,5 @@
 export { default as loadDrivers } from "./LoadDrivers";
-export { default as getUpScripts } from "./GetUpScripts";
+export { default as getUpScripts } from "./GetScripts";
 export { default as getMigrationsPath } from "./GetMigrationsPath";
 export { default as createScriptContext } from "./CreateScriptContext";
-export { default as getDownScripts } from "./GetDownScripts";
 export { default as processParams } from "./ProcessParams";
